+++ lib/hive_string_store.dart
@@ -1,15 +1,13 @@
 import 'dart:async';
 
-import 'package:hive/hive.dart';
-
+import 'hive_store.dart';
 import 'persistent_store_base.dart';
 
-class HiveStringStore<T> extends PersistentStore<T> {
+class HiveStringStore<T> extends PersistentStore<T> implements ResetableStore {
   final String Function(T object) encode;
   final T Function(String value) decode;
 
-  late final Box<String> box;
-  late FutureOr<T> Function() defaultValue;
+  final HiveStore<String> store;
 
   final String boxName;
 
@@ -17,30 +15,29 @@ class HiveStringStore<T> extends PersistentStore<T> {
     required this.encode,
     required this.decode,
     required this.boxName,
-  });
+    required Future<T> Function() defaultValue,
+  }) : store = HiveStore(
+          boxName: boxName,
+          defaultValue: () => defaultValue().then(encode),
+        );
 
   @override
-  Future<void> init() async {
-    box = await Hive.openBox<String>(boxName);
-  }
+  Future<void> init() async => store.init();
 
   @override
   Future<T> load() async {
-    if (box.isNotEmpty) {
-      return decode(box.getAt(0) as String);
-    } else {
-      final def = await defaultValue();
-      await box.add(encode(def));
-      return def;
-    }
+    return decode(await store.load());
   }
 
   @override
   Future<T> save(T newValue) async {
-    await box.putAt(0, encode(newValue));
+    await store.save(encode(newValue));
     return newValue;
   }
 
   @override
-  Future<void> close() => box.flush();
+  Future<void> close() => store.close();
+
+  @override
+  Future<void> reset() => store.reset();
 }