+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## [1.1.0] - WIP
 
 feat(refactor interface):
-- remove storeName, defaultValue from `init` signature
+- remove storeName, defaultValue from `init` signature and add it to store
 - resetable store and `reset` method for notifier
+- HiveJsonStore optional `toJson`
+
 ## [1.0.0] - 2022-08-04
 
 feat(initial):